[step3] auth, member 패키지 분리
