+++ src/main/java/nextstep/auth/application/TokenService.java
@@ -1,6 +1,8 @@
-package nextstep.member.application;
+package nextstep.auth.application;
 
-import nextstep.member.exception.AuthenticationException;
+import nextstep.auth.application.JwtTokenProvider;
+import nextstep.auth.exception.AuthenticationException;
+import nextstep.member.application.MemberService;
 import nextstep.member.application.dto.ApplicationTokenResponse;
 import nextstep.member.domain.Member;
 import org.springframework.stereotype.Service;
