+++ src/main/java/nextstep/auth/application/dto/OauthTokenRequest.java
@@ -1,4 +1,4 @@
-package nextstep.member.application.dto;
+package nextstep.auth.application.dto;
 
 import lombok.AccessLevel;
 import lombok.Getter;
