+++ src/main/java/nextstep/auth/application/dto/ResourceResponse.java
@@ -1,4 +1,4 @@
-package nextstep.member.application.dto.github;
+package nextstep.auth.application.dto;
 
 import lombok.AccessLevel;
 import lombok.Getter;
@@ -6,12 +6,12 @@ import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
-public class GithubProfileResponse {
+public class ResourceResponse {
 
     private String email;
-    private int age;
+    private Integer age;
 
-    public GithubProfileResponse(String email, int age) {
+    public ResourceResponse(String email, Integer age) {
         this.email = email;
         this.age = age;
     }
