+++ src/main/java/nextstep/member/application/MemberService.java
@@ -2,8 +2,8 @@ package nextstep.member.application;
 
 import nextstep.member.application.dto.MemberRequest;
 import nextstep.member.application.dto.MemberResponse;
-import nextstep.member.application.dto.ResourceResponse;
-import nextstep.member.domain.LoginMember;
+import nextstep.auth.application.dto.ResourceResponse;
+import nextstep.auth.domain.LoginMember;
 import nextstep.member.domain.Member;
 import nextstep.member.domain.MemberRepository;
 import org.springframework.stereotype.Service;
