+++ src/test/java/nextstep/member/acceptance/test/AuthGithubTestController.java
@@ -1,8 +1,8 @@
 package nextstep.member.acceptance.test;
 
-import nextstep.member.application.dto.github.GithubAccessTokenRequest;
-import nextstep.member.application.dto.github.GithubAccessTokenResponse;
-import nextstep.member.application.dto.github.GithubProfileResponse;
+import nextstep.auth.application.dto.github.GithubAccessTokenRequest;
+import nextstep.auth.application.dto.github.GithubAccessTokenResponse;
+import nextstep.auth.application.dto.github.GithubProfileResponse;
 import org.springframework.context.annotation.Profile;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
